New proc-macro
